Attempted Fix for Location without WiFi
Discovered that geolocation fails when no WiFi networks are present, such as when kayaking on a lake far from the shore. Attempting to fix this by forcing GPS usage with the high accuracy location option.
Cannot simulate this case in a test environment, hence labelling this as an "attempted fix".

Co-authored-by: Chris Bingham <3534465+binghamchris@users.noreply.github.com>